dev/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yukihito", email = "ok220020@oks.iput.ac.jp" }
]
dependencies = [
    "flask>=3.0.3",
    "pandas>=2.2.2",
    "seaborn>=0.13.2",
    "scipy>=1.14.1",
    "scikit-learn>=1.5.1",
    "ipykernel>=6.29.5",
    "google-generativeai>=0.7.2",
    "python-dotenv>=1.0.1",
    "flask-cors>=5.0.0",
    "chardet>=5.2.0",
    "colorama>=0.4.6",
    "ruff>=0.7.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.rye.scripts]
start = "python app.py"
